Gather denomination selectors into one file
