+++ src/actions/SendConfirmationActions.js
@@ -13,10 +13,10 @@ import { Airship, showError } from '../components/services/AirshipInstance.js'
 import { EXCHANGE_SCENE, FEE_ALERT_THRESHOLD, FIO_STR, PLUGIN_BUY, TRANSACTION_DETAILS } from '../constants/indexConstants'
 import s from '../locales/strings.js'
 import { addToFioAddressCache, recordSend } from '../modules/FioAddress/util'
-import { getExchangeDenomination as settingsGetExchangeDenomination } from '../modules/Settings/selectors.js'
 import { getAuthRequired, getSpendInfo, getSpendInfoWithoutState, getTransaction } from '../modules/UI/scenes/SendConfirmation/selectors'
 import { convertCurrencyFromExchangeRates, getExchangeRate } from '../modules/UI/selectors.js'
 import { type GuiMakeSpendInfo } from '../reducers/scenes/SendConfirmationReducer.js'
+import { getExchangeDenomination } from '../selectors/DenominationSelectors.js'
 import type { Dispatch, GetState } from '../types/reduxTypes.js'
 import { convertNativeToExchange } from '../util/utils'
 import * as UTILS from '../util/utils.js'
@@ -103,7 +103,7 @@ export const sendConfirmationUpdateTx =
         const insufficientFunds = asMaybeInsufficientFundsError(error)
         if (insufficientFunds != null && insufficientFunds.currencyCode != null && spendInfo.currencyCode !== insufficientFunds.currencyCode) {
           const { currencyCode, networkFee = '' } = insufficientFunds
-          const multiplier = settingsGetExchangeDenomination(state, currencyCode).multiplier
+          const multiplier = getExchangeDenomination(state, currencyCode).multiplier
           const amountString = UTILS.roundedFee(networkFee, 2, multiplier)
           const result = await Airship.show(bridge => (
             <ButtonsModal
@@ -160,7 +160,7 @@ export const updateMaxSpend =
         const guiWallet = wallets[walletId]
         const currencyCode = selectedCurrencyCode || state.ui.wallets.selectedCurrencyCode
         const isoFiatCurrencyCode = guiWallet.isoFiatCurrencyCode
-        const exchangeDenomination = settingsGetExchangeDenomination(state, currencyCode)
+        const exchangeDenomination = getExchangeDenomination(state, currencyCode)
 
         const exchangeAmount = convertNativeToExchange(exchangeDenomination.multiplier)(nativeAmount)
         const fiatPerCrypto = getExchangeRate(state, currencyCode, isoFiatCurrencyCode)
@@ -195,7 +195,7 @@ export const signBroadcastAndSave =
     const guiWallet = wallets[walletId || state.ui.wallets.selectedWalletId]
     const currencyCode = selectedCurrencyCode || state.ui.wallets.selectedCurrencyCode
     const isoFiatCurrencyCode = guiWallet.isoFiatCurrencyCode
-    const exchangeDenomination = settingsGetExchangeDenomination(state, currencyCode)
+    const exchangeDenomination = getExchangeDenomination(state, currencyCode)
 
     const exchangeAmount = convertNativeToExchange(exchangeDenomination.multiplier)(edgeUnsignedTransaction.nativeAmount)
     const fiatPerCrypto = getExchangeRate(state, currencyCode, isoFiatCurrencyCode).toString()
