+++ src/components/modals/AccelerateTxModel.js
@@ -9,8 +9,8 @@ import { connect } from 'react-redux'
 import { playSendSound } from '../../actions/SoundActions.js'
 import { TRANSACTION_DETAILS } from '../../constants/indexConstants'
 import s from '../../locales/strings.js'
-import { getDisplayDenomination } from '../../modules/Settings/selectors.js'
 import { Slider } from '../../modules/UI/components/Slider/Slider.js'
+import { getDisplayDenomination } from '../../selectors/DenominationSelectors.js'
 import { type RootState } from '../../types/reduxTypes.js'
 import type { GuiWallet } from '../../types/types.js'
 import { type GuiExchangeRates } from '../../types/types.js'
