+++ src/components/scenes/FioAddressRegisterSelectWalletScene.js
@@ -11,7 +11,7 @@ import { sprintf } from 'sprintf-js'
 import * as Constants from '../../constants/indexConstants'
 import s from '../../locales/strings.js'
 import { getRegInfo } from '../../modules/FioAddress/util'
-import { getDisplayDenomination, getExchangeDenomination } from '../../modules/Settings/selectors.js'
+import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors.js'
 import type { Dispatch } from '../../types/reduxTypes'
 import { type RootState } from '../../types/reduxTypes'
 import type { FioDomain, GuiWallet } from '../../types/types'
