+++ src/components/scenes/FioRequestConfirmationScene.js
@@ -11,9 +11,9 @@ import * as Constants from '../../constants/indexConstants'
 import { formatNumber } from '../../locales/intl.js'
 import s from '../../locales/strings.js'
 import { addToFioAddressCache, checkExpiredFioAddress, checkPubAddress } from '../../modules/FioAddress/util'
-import { getDisplayDenomination } from '../../modules/Settings/selectors.js'
 import { Slider } from '../../modules/UI/components/Slider/Slider'
-import { getExchangeRate, getPrimaryExchangeDenomination, getSelectedWallet } from '../../modules/UI/selectors.js'
+import { getExchangeRate, getSelectedWallet } from '../../modules/UI/selectors.js'
+import { getDisplayDenomination, getPrimaryExchangeDenomination } from '../../selectors/DenominationSelectors.js'
 import { type RootState } from '../../types/reduxTypes'
 import type { GuiCurrencyInfo, GuiDenomination, GuiWallet } from '../../types/types'
 import { emptyCurrencyInfo } from '../../types/types'
