+++ src/components/scenes/FioRequestListScene.js
@@ -15,8 +15,8 @@ import s from '../../locales/strings.js'
 import { addToFioAddressCache, cancelFioRequest, FIO_NO_BUNDLED_ERR_CODE } from '../../modules/FioAddress/util'
 import { FioRequestRowConnector as FioRequestRow } from '../../modules/FioRequest/components/FioRequestRow'
 import { isRejectedFioRequest, isSentFioRequest } from '../../modules/FioRequest/util'
-import { getExchangeDenomination } from '../../modules/Settings/selectors.js'
 import { Gradient } from '../../modules/UI/components/Gradient/Gradient.ui'
+import { getExchangeDenomination } from '../../selectors/DenominationSelectors.js'
 import type { Dispatch } from '../../types/reduxTypes'
 import { type RootState } from '../../types/reduxTypes'
 import type { FioRequest, GuiWallet } from '../../types/types'
