+++ src/components/scenes/RequestScene.js
@@ -17,8 +17,8 @@ import * as Constants from '../../constants/indexConstants'
 import { formatNumber } from '../../locales/intl.js'
 import s from '../../locales/strings.js'
 import { refreshAllFioAddresses } from '../../modules/FioAddress/action'
-import { getDisplayDenomination } from '../../modules/Settings/selectors.js'
-import { getExchangeRate, getPrimaryExchangeDenomination, getSelectedWallet } from '../../modules/UI/selectors.js'
+import { getExchangeRate, getSelectedWallet } from '../../modules/UI/selectors.js'
+import { getDisplayDenomination, getPrimaryExchangeDenomination } from '../../selectors/DenominationSelectors.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import type { GuiCurrencyInfo, GuiDenomination, GuiWallet } from '../../types/types.js'
 import { getCurrencyIcon } from '../../util/CurrencyInfoHelpers.js'
