+++ src/components/scenes/TransactionDetailsScene.js
@@ -15,8 +15,8 @@ import { getSubcategories, setNewSubcategory, setTransactionDetails } from '../.
 import * as Constants from '../../constants/indexConstants'
 import { formatNumber } from '../../locales/intl.js'
 import s from '../../locales/strings.js'
-import { getDisplayDenomination } from '../../modules/Settings/selectors.js'
 import { convertCurrencyFromExchangeRates, convertNativeToExchangeRateDenomination } from '../../modules/UI/selectors.js'
+import { getDisplayDenomination } from '../../selectors/DenominationSelectors.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import type { GuiContact, GuiWallet } from '../../types/types.js'
 import * as UTILS from '../../util/utils.js'
