+++ src/components/scenes/TransactionsExportScene.js
@@ -11,7 +11,7 @@ import { connect } from 'react-redux'
 import { exportTransactionsToCSV, exportTransactionsToQBO } from '../../actions/TransactionExportActions.js'
 import { formatDate } from '../../locales/intl.js'
 import s from '../../locales/strings'
-import { getDisplayDenomination } from '../../modules/Settings/selectors.js'
+import { getDisplayDenomination } from '../../selectors/DenominationSelectors.js'
 import { type RootState } from '../../types/reduxTypes.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { DateModal } from '../modals/DateModal.js'
