+++ src/components/themed/WalletListSortableRow.js
@@ -8,9 +8,9 @@ import { connect } from 'react-redux'
 
 import { formatNumberInput } from '../../locales/intl.js'
 import s from '../../locales/strings.js'
-import { getDisplayDenominationFromSettings } from '../../modules/Settings/selectors.js'
 import { calculateWalletFiatBalanceWithoutState } from '../../modules/UI/selectors.js'
 import { type SettingsState } from '../../reducers/scenes/SettingsReducer.js'
+import { getDisplayDenominationFromSettings } from '../../selectors/DenominationSelectors.js'
 import { type RootState } from '../../types/reduxTypes.js'
 import { type GuiExchangeRates, type GuiWallet } from '../../types/types.js'
 import { decimalOrZero, getFiatSymbol, truncateDecimals } from '../../util/utils'
