+++ src/modules/FioRequest/components/FioRequestRow.js
@@ -11,10 +11,10 @@ import { ClickableRow } from '../../../components/themed/ClickableRow'
 import { EdgeText } from '../../../components/themed/EdgeText'
 import { formatNumber, formatTime } from '../../../locales/intl.js'
 import s from '../../../locales/strings'
+import { getDisplayDenomination } from '../../../selectors/DenominationSelectors.js'
 import { type RootState } from '../../../types/reduxTypes'
 import { type FioRequest, type GuiWallet } from '../../../types/types'
 import { getFiatSymbol } from '../../../util/utils'
-import { getDisplayDenomination } from '../../Settings/selectors.js'
 import { getSelectedWallet } from '../../UI/selectors'
 import { isRejectedFioRequest, isSentFioRequest } from '../util'
